requirements.txt
flask
requests
pymongo
kafka_python